[DEL-4617]: Move redist metrics to 960 module to avoid circular dependencies
