+++ 420-delegate-service/src/main/java/io/harness/metrics/intfc/DelegateMetricsService.java
@@ -23,6 +23,4 @@ public interface DelegateMetricsService {
   void recordPerpetualTaskMetrics(String accountId, String perpetualTaskType, String metricName);
 
   void recordDelegateMetricsPerAccount(String accountId, String metricName);
-
-  void recordRedisSubscription(String topicName, double value);
 }
