+++ 960-persistence/src/main/java/io/harness/metrics/impl/PersistenceMetricsServiceImpl.java
@@ -9,35 +9,44 @@ package io.harness.metrics.impl;
 
 import io.harness.annotations.dev.HarnessTeam;
 import io.harness.annotations.dev.OwnedBy;
-import io.harness.iterator.PersistentIterable;
 import io.harness.metrics.beans.IteratorMetricContext;
+import io.harness.metrics.beans.RedisTopicContext;
 import io.harness.metrics.service.api.MetricService;
 
 import com.google.inject.Inject;
 import com.google.inject.Singleton;
 import java.time.Duration;
+import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 
 @Singleton
 @Slf4j
 @OwnedBy(HarnessTeam.DEL)
-public class IteratorMetricsServiceImpl {
+@RequiredArgsConstructor(onConstructor_ = @Inject)
+public class PersistenceMetricsServiceImpl {
   public static final String ITERATOR_ERROR = "iterator_error";
   public static final String ITERATOR_DELAY = "iterator_delay";
   public static final String ITERATOR_WORKING_ON_ENTITY = "iterator_working_on_entity";
   public static final String ITERATOR_PROCESSING_TIME = "iterator_processing_time";
+  public static final String REDIS_SUBSCRIPTION_CNT = "redis_subscription_cnt";
 
-  @Inject private MetricService metricService;
+  private final MetricService metricService;
 
-  public <T extends PersistentIterable> void recordIteratorMetrics(String iteratorName, String metricName) {
+  public void recordIteratorMetrics(String iteratorName, String metricName) {
     try (IteratorMetricContext ignore = new IteratorMetricContext(iteratorName)) {
       metricService.incCounter(metricName);
     }
   }
-  public <T extends PersistentIterable> void recordIteratorMetricsWithDuration(
+  public void recordIteratorMetricsWithDuration(
       String iteratorName, Duration duration, String metricName) {
     try (IteratorMetricContext ignore = new IteratorMetricContext(iteratorName)) {
       metricService.recordDuration(metricName, duration);
+    }
+  }
+
+  public void recordRedisMetric(final String metric, final String topicName, final double value) {
+    try (RedisTopicContext ignore = new RedisTopicContext(topicName)) {
+      metricService.recordMetric(metric, value);
     }
   }
 }
