+++ 960-persistence/src/main/resources/metrics/metricGroups/redis_metrics_group.yaml
